Cast int.

+++ st2/experiments/mnist/exp005.py
@@ -38,7 +38,7 @@ def main(args):
     learning_rate = 1. * 1e-3
     n_epoch = 300
     act = F.relu
-    iter_epoch = n_train_data / batch_size
+    iter_epoch = int(n_train_data / batch_size)
     n_iter = n_epoch * iter_epoch
     extension_module = args.context
